evm: add gasLimit to quoting functions [wip]
instead of the transceiver having to magically know how much gas to
quote for...

this is wip, because it's not backwards compatible. in order to make it
so, we could add functions that match the original signatures, and pass
0 to gasLimit, which will trigger the old behaviour of magicking up the
gas limit. that expectation should also be documented on the transceiver
side.