Prevent `rent_payer` from being overwritten if already initialized (#614)
Currently, it is possible to call `set_token_authority` twice consecutively - first time to initialize the `pending_token_authority` account and second time to update the `rent_payer`. This could result in the `rent_payer` being overwritten to the latest account instead of the actual rent payer. This PR fixes that by only setting `rent_payer` if `pending_token_authority` account is not previously initialized.