cli: check solana-cli version (#645)
* cli: check solana-cli version

This vastly improves the error message when the wrong solana-cli version
is installed. Before this change, it just prints a cryptic file not
found error.

Also changed the `checkAnchorVersion` function to read the expected
version from the Anchor.toml of the relevant worktree. This makes the
function robust in the future in case future ntt versions are upgraded
to newer anchor versions.

* Update cli/src/index.ts

Co-authored-by: Sriram Nurani Viswanathan <50625504+nvsriram@users.noreply.github.com>

* cli: switch solana version

detect the appropriate version picker and run it

---------

Co-authored-by: Sriram Nurani Viswanathan <50625504+nvsriram@users.noreply.github.com>