evm: split up transceivers into generic + ntt transceiver
This allows us to have a base transceiver contract which is agnostic to
the messages being sent. In particular, no token logic in these
transceivers. Then we inherit from these base contracts to add the
token-specific stuff. This way, the change is backwards compatible.

NOTE: I kept the contract names as is, so "Transceiver" is the
token-specific transceiver, and "GenericTransceiver" is the base contract.

However, there is a slight discrepancy now between the contract names
and the file names that contain those contracts. The specific stuff like
"Transceiver" is now contained in "NttTransceiver.sol". This is to make
git recognise that the original "Transceiver.sol" was renamed (and slightly
modified) into "GenericTransceiver.sol". Had we kept "Transceiver.sol"
for "NttTransceiver.sol", git would show a much larger diff, effectively
comparing the "wrong" files.

We can revisit this later.